Refactor and accept default values when annotations are not complete. In the same way as annotation processor
